Release a button and the hero moves! Create the updatePlayer method and implemented the drawPlayer method; I have not have written the code persay, so I should at least describe what it's doing:
the way it works is by using double buffering, which is where the image
is drawn in memory and then displayed when complete; this is to prevent
the screen from updating before the drawing is finished. In the code,
the player/mask are drawn on the hdcBuffer using the BitBlt method, and
then once they're composed BitBlt is used to draw the hdcBuffer on the
main hdc window.