Ok, couple of changes!
So the most important one is the change in object hierarchy; 
The object struct is stored in its own header (top level). from there,
it's included by a header for public functions (to be used by other
sources) and another header which has all functions (to be used for
testing public/private functions). Then there's the associated .c file
which will contain the actual functions, which may follow the form
verb<objectname>() for clarity's sake. Also, this will help leave the
door open for things like polymorphism and such. 

I've implemented the new object (character) in the main window; it's
basically taking over the playerInfo object, but it works! The next step
on that front is to move the HBITMAP piece into the character object as
well.

I've also set up the testing framework for the character object, as well
as testing pretty much the only method (nameCharacter)