committing what I have for now since it's been a while.
Introducing the field class. this has two structs; field and
characterArray.  basically I'm setting it up so that it's a 3d array; x
and y position on the field (double array of characterArray) and for
each position, they have z characters on the space. I've set the upper
bounds for the field of 25x25, since that's likely as big as I'll ever
need it to be (since 25*50 pixels = 1250 pixels, plenty huge). Now I'm
in the process of creating some meaningful get/set methods for
characters in the field... not entirely sure how that dynamics going to
work yet. Anyway, we'll see.