Still impementing the attack, need to work on that.
I think I was initializing the field wrong because there were some
issues setting a individual, didn't properly malloc the objects. Also I
wasn't able to properly set the skeleton to a space, so the first thing
I did was move the skeleton, which worked. Will fix that in the future.

I also fixed the move cursor object so that it moves in all directions,
and moves similar to an individual.