I realize it's been a while since I've committed, so I'm going to take a snapshot now in case something terrible happens.
Anyway, after kicking it down the trail long enough, I've begun working
on the pathfinding algorithm. It's actually going fairly smoothly, and
I'm even learning quite a bit about how to not do things stupidly.

For example: I was using malloc(), which returns space for an object in
memory, whereas I can just have a pointer point to a pointer of the
object in memory, and deref it to use it-thus freeing up a lot of
unnecessary copying/mem-usage. I guess I've just been doing a lot of
overkill.

Anyway the pathfinding algorithm is mostly done, I just have to
implement the calling function and finish up a bit of the main recursing
method.