HAPPY BIRTHDAY TO ME, I FIGURED IT OUT.
So I found my problem, I was being an idiot; I was trying to implement
and array of pointers to pointers, but what I actually needed was just
an array of pointers. Once I updated that and tweaked a few things the
algorithm worked! I'm going to commit it now and then implement it. 

Horay!