Usually you look back on code you wrote 3 months ago and think, "what was I thinking?", for me, I'm amazed at the code I write the day before.
so the cursorMode / moveMode stuff was really dumb and I fixed them to
just be ints, and update it only when it left the main window as a
pointer (I had realized at this point that pointing to the same variable
with two different pointers updates both of them, which now seems
obvious, I know). 

Anyway, I fixed that, changed moveNode to not be an array (yeah, I know)
and implemented more of the move loop. I also found a much smarter way
to handle the keyboard input; basically if you mod the hex value by 16,
you'll get the numerical value (quite a nice design choice). I'll say it
now that it's not tested yet, so we'll see if my next commit debunks
this. Getting closer to having a sweet movement system!