Added new tests for attack, changed method name to have proper spelling (wont be the last, im sure)
2 other interesting things happened; when I hit Ignore on one of the
asserts failed popup, that ignored them forever and crashed the program.
I was able to fix this by creating a new workspace - and in doing so, I
tested my first code pull. Not only was I surprised everything worked,
but it even saved my eclipse configurations to run the project.

Aside from that, it's been a busy week (I haven't done much because I've
been coding at work for 11 hours) hopefully after that stuff's done I'll
have more time to work on this.