Alright, I'm back baby.
So after a week or so of not touching this project, I've sprung into
action by creating the field_controller, which contains an array of
enemies--that's right, now we got a whole ruckus of enemies! complete
with adding/removing functionality, this is definitely a milestone :)

I also updated cursorloop so it has a post-cursor loop to deal with
player actions; I can probably make a general flag to handle that.

One minor bug is that the individual is not removed from the field when
they die; but aside from that, combat is definitely coming together!