The status bar displays the player's health... in real time!! super cool! I'll be adding a plethora of information to that!
A lot of the stuff I did here was a bit of a pain; mostly dealing with
the win32 api side of things. For example, to update the window requires
GetDC(), however in the wrong place GetDC can suck up all the processing
and block other windows! lots of little things to keep track of. As for
the structure of the sidebar_window.c, im not entirely sure how I want
to go about it; it may just end up as a nasty block of junk to specify
the TextOut calls... we'll see