I went bananas and implemented a bunch of individual/attack related stuff.
I added AC, attack, and critType attributes to the individual, changed
the attack system so that the individual actually has to do an attack
roll, do the crit type if the roll is a 20, and attack or miss depending
on the roll.

downside is I had to deactivate the tests for now, and I'll have to
revamp them to fit the changes; at this point I may as well implement
weapon types and DR, but I have to tackle magic as well. I should make a
note to get a better understanding of that crap this weekend (going to
Madison, woot).