I fixed the path finding algorithm so that individuals can 'walk through' each other onto a space in their path that isn't occupied. This fixes the issue of them being dependant on move order, however it doesn't solve the issue of them being in one long line with open spaces around them to move to.
The technical changes I made was to add another int to the nodeArr
struct which would keep track of the length to the closest-open-spot of
their move path, constrained by the length/move distance of the
individual.

Space was also modified to have a 'tmpIndividual', exclusively used for
animations, to hold individuals as they moved from space to space. This
was needed because the individual should move from one space to the
next, and the current struct didn't have the capacity for more than one
individual on a space at a given time.

what do you call an alligator in a vest? an investigator.