AH HA!! AH HA, AH HA, AH HA!!*
So I've been working on this on and off for the last few weeks (I did
take a break to play a game [dark souls]), but I had been working on
making the pathing algorithm more robust. The reason for the "AH-HA"-ing
was I had an issue where two individuals would move on top of each
other, but in the debugger, they would work fine (truely an awful
problem). Anyway after reading up on why that happens, I found a Stack
Overflow question which basically said  "The Debug will work when the
program invokes undefined behaviour... the Heap is laid out differently
- debug-mode allocators may pad the allocations with extra space to
catch bugs" (thank you, nneonneo). In short, the problem was the new
field I created in the field struct (tmpIndividual - used for having a
tmp individual on a space), so  I commented it out and now it works (I
suspect it's because I didn't initialize it properly, will fix in the
future).

Pathing Algorithm update: the algorithm itself hasn't changed, meaning
it still constructs a path to the player - however there is a
post-processing step which guarantees the new space is Open and Within
Movement Range; basically, creating the path to the individual insures
that the enemy is going the right way, however the must make a new path
that's based on how far they can traverse on the current path (e.g. if
the path-to-player is 10 spaces but the enemy can only move 3, then a
new path is created based on the [3]rd space in the path-to-player path.
Then, it tries to find the next open space (i.e, unblocked by
individuals or unpassable terrain); if [3] is open, enemy goes to that
space - however, if it's taken, it goes to space [2], but can move burst
1 (because if they can move 3, any of the spaces surrounding [2] would
be accessible (Note there is a need to sort the spaces so that the ones
closest to the front of the path are considered first)).
An important note is that the enemy cannot walk through the player; this
means that if the enemy can move 11 and the path is 10 long, they will
not have there extra movement carry over. This is because you cannot
garentee there is a way around the player (e.g. they're standing between
2 unpassable walls, but the space behind them is open - it would be
illegal for the enemy to pass through the player to the space behind
them.

*whew* what a commit, I gotta make eggs for the wife and I and get to
work!

*sorry, the endorphins just wore off from solving the problem--man that
was buggin me.