I've done something I hadn't accomplished in prior endevors; I've made it so bitmaps can be rotated! this will be extremely useful for spaces where I can simply rotate a piece to create borders (e.g. the shoal of a river) without creating an exhaustive number of images - an imperitive task for creating a game!
Currently I just have it with some test images in the field where the
background is drawn, but maybe I'll port this as a more generic method
and include it for individuals and such who are lying down (or slayn).

As for what I've learned, xform and changing the global bitblt form is
kinda crazy; when the image is rotated, the entire HDC is rotated with
it; which (in some cases) swaps the coordinates... quite an interesting
problem to overcome! I even had to dust off some of my math skills with
radians and such.