I added background spaces which use the new rotation functionality, and I updated the way map data is stored and read; now, it reads two characters-the first being the key for the space, then the next being an arrow (^>v<) to determin the orientation of the object.
Also while doing this, I found that the 25x25 field wouldn't cut it for
my worldbuilding, so I knocked it up to 100x100, I've had no issues
running it (im a little more concerned what 1000x1000 will be like, for
the overworld) 

I also spent quite a bit of time working with how the shore backgrounds
look; I really liked the waves I had put on the sea spaces, however it
wouldn't be possible to have the shores replicate the horizonal waves
because they could be rotated (and therefore, verticle waves). I decided
the best course of action was to make it so the waves were approaching
the shore (which makes sense, right?). It looks ok for now.

Man, it's late.