fixed a bug after I made a change to movepath; I was wondering why no matter how far away, each skeleton would move; this was because I had changed the move limit to 100 instead of 10. Consequently, I had not tested for a case when the enemy was out of rang, and as a result it would crash (when I returned a null instance, I had not created the result nodeArr appropreatly). Anyways, it works now.
Not a coding thing, but I removed the waves from the sea images all
together, it just doesn't go with the rest of the map I feel. Also
there's a conflict of tones--the grass background is very basic, the sea
stuff was cluttered and chaotic. It also seemed relitively zoomed in,
which would contrast if I put in mountains; we'll see how it turns out
when I add a world map and a local map, at the very least I can reuse
these art assets.