I've surpassed another milestone - the player can now travel between two maps!
The system I devised, as aluded to in the previous commit, uses the map
file to link to another file (second line, formatted to
<map_n>_transit.txt). For each line in that file, it is formatted as the
transit ID, x, y, name of the map being transited to, and that maps
transit ID. The [x,y] links the ID, targetID and Target Map to a space.
When the user presses 'e' on a space, if the target ID is not zero
(default), then the user is warped to the target map, using the [x,y]
coordinates from the target ID (the target ID will exist in the target
map).

This change has a few reaching effects I have to deal with, from
changing the shift to having some way to save enemy data  when the user
leaves and comes back (do they heal? do they stay dead for the last n
rooms?) However for the moment, I'm just going to enjoy what I've got.