C is a strange language.
After having the tests disabled for a while, I decided it was about time
I fix them, only to have them barf from the get-go. I would make the
player, the enemies list, the field... but it would error when I wasn't
in debug mode (which means that some undefined behavior was happening).
In the classic c way-adding a print statement caused my code to work,
and when it was removed, it didn't.

wow.

After some googling, I found a few SO posts saying how it was due to
undefined behavior, and that the printf() was changing how the program
was working just enough to make it work-although this is a great, big
red flag on I've done something terribly wrong. 

What I think the issue was, was that the individual class had a name
field defined as char *, and that was throwing off the sizing of the
individual. Since I was already mallocing the size of name to be 32
default, I converted it to char[32] and it appears to be working.

That terrifies me. 

It's like I joined the military, and I'm sitting in the jungle on watch
over my sleeping platoon, and I heard a twig snap in the distance. I
dont know what's out there, but I feel that this is what I signed up
for, and...

I'm going to bed.