Ok, NOW I think it's working.
Last time, I had cleverly disabled the tests, then assumed everything
passed insted of verifying it was actually running! Fortunately, I was
on the right track in terms of figuring out what was wrong, and it was
like what was happening last time - the transitData struct had a
datatype of char* for the map name, and I converted it to char[256],
thereby stablizing the size of the object, and causing everything to run
correctly. I'm going to investigate further into that, but again, it's a
lovecraftian horror - how do I know it will never happen again? How do I
know that the grotesque heisenbug beast is not mere inches beneath me as
I continue my c expedition?