Updated the tests some more, there's a test for skeleton5's path finding both for direct and truncated, as well as when there's other skeletons in the way and it must find another space to go to. Also tested attacking mechanics.
The next step was to test the map loading functionality when the player
goes to different rooms, however when I tried moving some code into its
own method, I encountered a strange bug-basically, the map would load
alright, but after 5 enters/exits, it would result in a blank screen,
but not crashing. I believe this to be a memory management issue with
mallocing pointers, but my attempts to fix it changed several files over
a few days of working on it, so I decided to roll back and just commit
my tests.