I've developed an in-game console and how it will be used. Two new structs: Console, which is a global object in the console.c source and is used indirectly with console-write calls, and messageNodes, which contain a char[256] message, and a link to previous and next messageNodes. The console object keeps track of the newest message node, as well as the current (which will be used for when the player cycles up/down between previous messages.
Some technical notes - messages are displayed from bottom going up, and
when messages span (a currently arbitrary) length, they are indented and
continued on the next line. The entire message is desplayed by printing
'chunks' of it from end to beginning, chaning the draw height each time
a line is printed.

The code is in need of a good refactoring, but aside from that, it's
coming along nicely. Future additions will include the player selecting
answers to questions, typically in npc dialogue.

yay.