I've finally implemented a hyphen system, measuring the chars in pixels, and having the text scroll with the window.
Measuring the chars in pixels: this was done using a win32 function that
takes in a string (and some other stuff) and returns the width in pixels
of that character. The row length is in pixels, and the counter is
incremented by the width of each character, thereby giving a more
accurate reading.

The way the 'hyphen system' works is when creating the spaceIndexArr, if
the most recent space character is 3 indexes away from the current index
(i), then the break will be inserted there. If not, then the current
index is used. later in the draw function, a check is made to determine
if the character at the index is a space or not (excluding first and
last indexes). If it's not a space, it appends a hyphen to the end of
it.

Lastly, leveraging the window rectangle, I adjust the left indent and
row size by taking (prc->right - prc->right * frac), where 'frac' is
some fraction of the row. this way it scales properly when the window
grows or shrinks, with all the text staying inside the window.

NOW--this 'iteration' is drawing to a close, having fulfilled all my
previous goals (field moves with player, can go to other fields,
implement a console, and displaying info to the user by window). I'll be
updating the Readme to include the goals for the next iteration: there
are 3-4 iterations left, I think: Implementing items, NPCs (that's gonna
suck), the special magic-system, and story building and misc. I'm
thinking that next iteration it's going to be items, and some other
important stuff I haven't gotten to yet.