Finally - my first philosophical game feature!
I've introduced 'existence', meaning whether something in the game
exists or not is kept track of by the things ID and a new global
registry.

global registry: This struct will be what keeps track of the game state.
Right now it keeps track of what exists: Currently it uses an int array
that can store up to 32,000 entries (on a 32 bit system, I've taken into
account that it could be 16,000 or otherwise). The int array is rather
cleverly exploited-since I only need a bit to say if a given ID exists,
and each int contains 32 bits, a 1000 int array can be used to store and
keep track of 32000 bits (I used
http://www.mathcs.emory.edu/~cheung/Courses/255/Syllabus/1-C-intro/bit-array.html
as a reference).

for example, if I have an object with ID 4 and I wanted to update that
bit, since 4 is less than 32 it would be the 0th element of the array,
and it would modify the 4th bit (via bitshifting). I would then check or
modify that bit to determine if the thing exists.

Pretty Effin' neat.