SO: I've made it so an item can be created from a file, and equipted to an individual.
When I write that out, I realize that it's not terribly impactful,
however it was quite an ordeal; mainly because there's so much crap to
keep track of with an item and dealing with file i/o. however, skeleton
one now wields a dagger. It may not be changing the fight (i.e. I
haven't implemented how items affect individuals yet) but it's certainly
a step in the correct direction.

I've had to update the file structure accordingly; the main map file now
requires another line in the beginning specifying the enemy items, and I
imagine I'll be doing it again when implementing on-map items.

im tired.