Implemented items in combat, added a few helper methods (add/remove/equip item), and created a new struct to hold the inventory stuff, attached to an individual (hence the variable name 'backpack' :) )
Some higher-level decisions made were to remove the 'weapon' and 'armor'
types, and instead have a 'type' variable associated with an item - this
can leave the door open for future implementations (e.g. rings, amulets,
etc.) and treats every item as a generic item, and checks the type for
specific things; for example if they type is 'w' (weapon) then it has a
meaningful weaponDamType value.