The field now has a dagger on it! I've created a fieldInventory structure which contains the items and total, and also it reads data from a file to create the items.
 I encountered some strange bugs, which happened when I accidentally
deleted a loadMap function call and stuck some initialization stuff in
the wrong place (ugh).
 
 Next up is some new functionality to implement (g)etting items from the
map which are on the ground, from there managing a way to print the info
in the console and select an item. There's also some other one-off stuff
where more data needs to be processed from the items (e.g. min/max
damage mod)