The player can now get items from the field and have them in their inventory, while removing the item from the field via global register. Also, max/min damage mods from items are taken into effect when damaging individuals.
I've drawn out the next step in the process, which is the inventory
view. When the user hits a key ('i', pobably) they will launch a new
loop. This loop will draw a box to the screen, which will list the items
the user has. the user will be able to scroll through the items and
view/use/equpt/unequpt the selected item. This is a tall order, but it
will borrow some similar functionality from the console view.