Players can now equipt/unequip items from the menu! The menu also reflects these changes by printing out (E) next to the item name.
The process for equiping goes as such:
player hits enter, the selected item is retrieved and sent to the
individual.c sourcel to be updated. A switch is done on the item type,
and for the case of weapons/armor. A check is made if the item is
currently equipt - if it is, it is unequipt. If it's not equipt, a check
is done on whether another item is equpt, and if not, the item is equpt. 
This is nice because it allows for further item-logic in the future
(e.g. can the player wield 2 weapons? can they wear this type of armor?
etc.)

The next task will be implementing logic for items - this will require
some thought.