Items are now consumeable! This is only for a small case (e.g. if an item restores health), however the infrastructure is in place for expanding with more things.
If the player consumes an item, the item will be removed from their
inventory, and the viewedItems array (what's displayed in the inventory
menu) is refreshed, the selector staying in place if possible, or
shifting down as needed.

I also fixed a bug where the inventory was being improperly iterated
over (i.e. using the count of the elements and not checking for NULL).

So, the player can now heal themselves with the excessive amount of
herbs surrounding them! Makes taking on the skeleton horde a bit easier.