Two major changes: Duration items now work, and the process to get an attribute has been changed.
Duration Items: These Items are type 'i', with itemType 'd'. They use
the minTurns/maxTurns values in the item struct to determine effect
lenght. at the start of each turn, the item is 'consumed', meaning it is
used for stats which will be updated and frozen (e.g. hp and mana). The
turn counter is decremented at the start of each turn as well, and once
it hits zero the item is removed. In the mean time, the temp item
attributes are factored into all attribute checks using the method
below.

Attribute Checks: Now, given a individual and a string that corrosponds
to a particular attribute (e.g. 'attack', 'ac', 'slashDR'), a number
will be return which factors in the individuals base stats, their equipt
items, and active items. This method (although a bit big) makes it quite
easy to deal with current attribute checking, and it makes it much easer
to expand upon.

words words words. Time to write tests.