One feature change- added a description value to items. If the user hits (d)escription in the inventory menu, the description will be printed in the console (I didn't feel like making another window view with formatted text, however I may come back to that).
The other thing I've been working on is data leaks - I downloaded a
program called Dr Memory (similar to valgrind on linux) and It reported
a crapton of interesting things going on. I know there are several
problems with the pathfinding algorithm, it may require a re-write in
the future, but lets see if I can patch it up in the meantime.