Two developments: 1) (I think) I fixed a major memory leak issue; specifaclly the currentIndividual and transitInfo pointers in the 'space' struct. I had been malloc-ing both values during initField() despite not requiring either values be populated. Now, when trasitioning between two fields, there is no or little committed memory increase (based on Resource Monitor)
2) I've made it so that the directory paths need not be fully qualified
(for loading .txt files). This is good because now I can develop on my
laptop without hastle! However if I want to drag the executable to
DrMemory, I'll have to update the paths to be absolute. Really the long
term fix is to have a properties file, but it works for now.

Also, this iteration is pretty well wrapped up, so I'll be working on
laying out the next one... also wow, record timing on this one!