Wow, I had no idea how malloc worked when I wrote the path finding algorithm... of course, I'm sure everything I'm writing today will make me cringe in 6 months anyway (this is the mark of a good programmer, I'm told). That said, I am greatful for the test suite I've created thus far, it's making the refactoring as painless as possible.
Anyway, I fixed all the leaks in the path finding algorithm, which
required a few little restructures of the code. The biggest memory
offences were re-malloc-ing the allnode and activenode arrays EVERY
TIME, in several locations!

Also I added a little feature for commenting out
enemies/items/transitInfo by placing a '#' in the first character of the
line of the respective data files - this makes it much nicer to edit.

Gobble gobble.