Tentative Commit: So shame on me, the last commit I did broke the tests, but I hadn't checked before I uncommented them... anyway, there was a unique issue with destroying the nodeArr, specifically that a node was referenced both in the nodeArr and the activeNodes/allNodes arrays. After calling free on the allNodes/activeNodes arrays, it would inadvertently free some of the nodes in nodeArr, causing a segfault!
 I decided to fix this by cloning the nodeArr values as needed -
however, I still want to clean it up. I was getting a lot of grief
because I was trying to make a double pointer to reference the nodes as
I cloned them, which totally works in the initEnemyActionMode loop, but
segfaulted every time. I somewhat cleverly converted it to a node *
array of size 1, and it worked wonderfully (although kind of silly). I
want to address this and make it a little better.
 
 Also there's a new small leak, which was the whole reason I did this.
 
 ...