OK. NOW IT WORKS.
Yeah so for that last commit, when I ran it in debug mode it kept
segfaulting, and I realized that it was because I was focusing on the
inital node, not all the previous nodes attached to it.

So for this solution, I used a new tactic: basically use the endnode for
the algorithm, but make clones to stick in allNodes. Also, since ~all~
the nodes get added to allNodes, there isn't a point in activeNodes
getting freed, since they are completely redundant for allNodes (also
caused segfaults, node got freed in active, faulted in all). So with
this new tactic, new shallow nodes are created (used only for their x/y
component, not the linking) and the nodeArr can be destroyed in peace,
along with allNodes -- no collisions.

I checked that the tests were working 2 times. Also, no leaks.
Tomorrow's a new day, I look forward to moving on from this node saga.