moving enemies are now drawn in front of others as they go. I've tested out my changes with a large group, and everything worked properly.
bit of a dilemma that I believe I've solved; currently, as an individual
moves, they are not connected to the field. As such, nothing can
interrupt their movement or they may end up on the same spot as another
individual. This is fine for most things, however something like traps
have to have strictly defined rules: if an individual steps on a space
that has a trap which they can trigger, it either must go off
successfully, or it must be disabled, allowing free passage. It doesn't
make sense for a trap to trigger potentially several times, because it
could stop an individual on the same space as another individual, and a
nearby open space is not guaranteed. The only time where a trap may work
multiple times is if it unavoidably insta-kills individuals (thus it
would free up the individual every time)

in short, there can be no circumstance where an individual is stopped
mid-movement, unless it's guaranteed that there is no individual there.