Hope you brought your snorkel, cause we're going overboard!
After much deliberation, I've decided to postpone events, because events
largely hinge on objects which are not loaded in memory. Where I'm going
with this obviously is I'm going to load EVERYTHING in memory! 

At first I was incredulous at how much more memory it would consume,
however an array of 1000 individuals and 5000 items pushes my program's
memory consumption up 1MB, which is a drop in the bucket. Further, the
maps would unload and reload as needed.

The main reason why I'm doing this is because my program lacks the
ability to keep track of something when it's unloaded. This makes
handling events very confusing, accident prone, and something ripe for
poor design. More over, I will gain two things: One, instead of loading
the same properties from the file every time, individuals/items will
pick up where they leave off in their same state, which leads to Two, it
will make saving the game significantly easier.

Now this data requires responsibility: Items and Individuals will only
be modified when they're active (and apart of npcs/enemies) or when an
event triggers a specific change. There wont be any willy-nilly changing
otherwise (lest I come up with a very compelling reason at the time).

I've made the infrastructure for this in global registry, which will be
converted to a static instance. At the start of the game, all
individuals and items will be loaded into their registry, and pulled
from when loading groups etc.