During my week of sickness, I've been working on how to best design event handling and I've finally come upon a viable solution.
 To trigger an event, there has to be a function call inside of the
action function - for example: to trigger an event on an individuals
death, there must be a check (triggerEventOnDeath()) when the individual
has received a mortal blow.  The check will need an identifier (in this
case, the individuals id) to see if there is a matching event ID in a
mapping for this specific event (onDeathIndividualEventMap). If the
event exists, it calls a function in main_window, passing along the
event ID, to process the event with the functional variables player,
npcs, enemies, and main_field.
 
 The reason I added a function to main_window even though it's usually
against my practices is that A) it reduces a tremendous amount of
complexity and B) functions without access to player/field/etc. can gain
it exclusively for the purpose of events.
 
 Now it's a matter of creating some events to process and loading them
into the registry and handlers