I've added STATs to the individual (including LUCK), and updated it so that health is correctly based off of CON, and Magic based off WILL (while I renamed the incoming totalHealth and totalMana parameters to baseX, which will be the dice spreads). Other things, such as AC, DAM, etc. will be calculated ad-hoc.
I also updated the sidebar to reflect the actual stats of the player,
not just 0s. It also marks the first ocasion where I used an inline if
statement, because it would have been tedious, and I think perhaps more
daunting to expand out that tiny piece of code (if the number is
negitive, shift slightly to the left - that's it).

It's a little unfortunate that I didn't build the individual with STATs
right away, that would have saved me a bit of pain. Perhaps it's just
the idea of what the game's suppose to be is morphing in my head.