More progress - I implemented a new item parameter to keep track of the weapons STR modifier, I also fixed the tests and a bug with creating a simpleDialog.
I did encounter some annoying issues while doing this; atoi() returns a
int, guess what atof() returns? that's right, a double. I was assigning
it to a float, which was creating some bizarre issues when I was passing
it in as a parameter--it would shove all the other parameters 'down one'
so to speak, and strangely it didn't cause everything to crash. C is a
wonderful tool which can do a lot, but has very few guards.