Well you learn something new every day - today I learned that strtok is a crappy function, because it cannot be performed more than once at a time on different strings. Evidently, it uses a global variable to keep track of which string it's tokenizing, so if you interrupt it with another, it will change that variable and ruin everything.
Today I welcome Charlie Gorden's strtok_r function (one I may re-write
for learning's sake). It takes in a character pointer which stores the
remaining string, and is used everytime strtok_r is called for that
particular string. Quite handy - and it fixes my delema with loading the
Effect (which is now working).