Created a function to calculate mana cost - with it, it introduced some more business logic that I hadn't accounted for. Fortunately at this point, I feel that I have a good amount accounted for, and that I can continue on.
At this point, I see the following major tasks ahead of me: 1) create a
view for the user to create an ability, 2) implement the ability logic
in the various locations, 3) allow the user to trigger an ability.
Example cases could be 1) player creates a fireball ability, 2) the
mechanic of the fireball are set up, 3) player uses the fireball on an
enemy.