total overhaul! I completly changed how the name is entered, and in doing so I created a new interface for getting text from the user. The interface is similar to what's used in many games for text entry using a D-pad (e.g. Legend of Zelda). I decided to make the change after realizing the system I had in place worked poorly.
The interface include a new image which has all the necessairy text
choices, a loop function, and various helper functions to interpret the
user input. since it's a standalone loop, other loops which implement it
(e.g. ability creation) must go into a waitForNameMode, and when the
name mode completes (either by name entry or escaping) the reciving mode
must interpret the value of the name mode. 

I still need to add functionality for backspacing and finishing name
entry through the GUI, however most of it is finished.