Well, it's about time to get the big ball rollin. I'm starting to implement abilities into the game mechanics, starting with the concept of 'activeAbilities', and a function similar to the existing getAttribute() ones. Also, in doing so, I realized that I've neglected the 'attack' effect from abilities, so I've implemented it in record time. Aside from that, some minor tweaks to the name box.
Also, I've recently made an important decision -- which I'll try not to
read to deeply into so I don't psych myself out -- I made the repository
private, because I have adopted the goal of getting this game in steam.
It'll require a lot of polish, but the bar seems a bit lower than
expected over there - that said, I wont submit it until I feel
comfortable with it.

Now the real quest starts!