As an added requirement to abilities, Ive begun working on status effects. Pretty standard, a individual can have up to 16 affects (poison, confusion, burning, paralysis, etc.) which are checked at the beginning of each turn and during specific actions.
I've created a container struct and implemented a standard processing
method. I'll shortly be able to implement diceDamageDuration from
targeted and self abilities.

One thing that I think I may have overlooked in abilities, and as
something I'll have to handle here, is that an individual may die as the
result of a status or ability, and that has to be caught and processed -
I've made a note to add that functionality in the future (it may be a
deeper-rooted issue)

Also, I forgot to commit the abilities_view background, so it's in this
one.