Major refactoring with individualGroups! So way back when I concluded that I could treat the list of enemies as a unsorted, self balancing list, I was incorrect.
Take for example the following senario: an enemy casts an AOE ability
and it ends up killing some enemies. This means that there would be
looping through A) enemy actions and B) enemies impacted by an AOE
ability. If an enemy is killed and removed from the group in B), the
list would be rebalanced, and A) wouldn't evaluate individuals which
were used to 'plug the holes' in the list

So, to resolve this, the autobalancing ability was removed. This means
that every iteration over a group goes from 0 -> MAX_INDIVIDUALS. Order
is kept, so that individuals wont be skipped, however if an individual
is removed there will be a NULL in the array.

Before this becomes way to long, I've also decided against having static
individualGroups.