I can say with high confidence that permanent ability functionality is implemented!
Over the course of accomplishing that, I had to do a few notable
refactorings, the big one being getting rid of 'totalhp' and
'totalMana'. The 'total' attributes are actually composed of other
attributes (baseHP and CON, baseMana and WILL), and should be calculated
as such whenever used-this is a rule I'd like to keep with future
composed abilities as well.

Also in a similar vein, getAttributeSum no longer acknowledges hp and
mana requests from items or abilities. This is because they're a special
case, and are really atomic to the individual (static).
 
 To my surprise, I noticed that I haven't implemented the weapon
Strength modifier yet - I'll do that before proceeding to the next
ability type.