Getting around to adding tests. So far, added one ability test for perminant abilities, so far so good.
One thing I noticed was that the selectedType on abilityCrationInstance
is set indirectly through the draw method - this strikes me as bad
design, but there really isn't a great alternative in this senario, as
the selected ability is driven by user input. I recoreded a bug to at
least set a default for this or a null check.